restructure the way platform references are set
in seriesParameters as PlatformRepository cannot be
autowired within SessionAwareRepository (itself does
inherit from SessionAwareRepository) which caused
`NoSuchBeanDefinitionException`s during autowiring
