+++ spi-impl/series/series-dao/src/main/java/org/n52/series/db/da/AbstractDataRepository.java
@@ -43,9 +43,13 @@ import org.n52.series.db.beans.parameter.Parameter;
 import org.n52.series.db.dao.DataDao;
 import org.n52.series.db.dao.DatasetDao;
 import org.n52.series.db.dao.DbQuery;
+import org.springframework.beans.factory.annotation.Autowired;
 
 public abstract class AbstractDataRepository<D extends Data<?>, DSE extends DatasetEntity<?>, DE extends DataEntity<?>, V extends AbstractValue<?>>
         extends SessionAwareRepository implements DataRepository<DSE, V> {
+
+    @Autowired
+    private PlatformRepository platformRepository;
 
     @Override
     public Data<?> getData(String seriesId, DbQuery dbQuery) throws DataAccessException {
@@ -54,7 +58,7 @@ public abstract class AbstractDataRepository<D extends Data<?>, DSE extends Data
             DatasetDao<DSE> seriesDao = getSeriesDao(session);
             String id = DatasetType.extractId(seriesId);
             DSE series = seriesDao.getInstance(id, dbQuery);
-            series.setPlatform(getPlatformEntity(series, dbQuery, session));
+            series.setPlatform(platformRepository.getPlatformEntity(series, dbQuery, session));
             return dbQuery.isExpanded()
                 ? assembleDataWithReferenceValues(series, dbQuery, session)
                 : assembleData(series, dbQuery, session);
